Extract donations with tokens behavior
